Factor out project slice
