+++ src/inmanta/server/__init__.py
@@ -25,3 +25,4 @@ SLICE_DATABASE = "core.database"
 SLICE_TRANSPORT = "core.transport"
 SLICE_COMPILER = "core.compiler"
 SLICE_FORM = "core.forms"
+SLICE_PROJECT_ENV = "core.project"
